requirements.txt
python-multipart==0.0.9
python-decouple==3.8
Authlib==1.2.0
fastapi==0.110.1
PyJWT==2.8.0
uvicorn==0.29.0
httpx==0.26.0
itsdangerous==2.1.2
PyYAML==6.0.1
pymssql==2.2.11